Skip remote by default

+++ git-clean.sh
@@ -1,9 +1,9 @@
 #!/usr/bin/env bash
 
 if [ -z "$1" ]
- then
+then
     BRANCH=origin/develop
- else
+else
     BRANCH=$1
 fi
 
@@ -12,6 +12,12 @@ COMMIT=$(git show --format=%H ${BRANCH})
 git fetch --prune
 git branch --merged | egrep -v "(^\*|master|develop|release|HEAD)" | xargs git branch -d
 # push to remote
-WEEK_AGO=$(python -c "from datetime import date, timedelta; print(date.today()-timedelta(days=7))")
-git branch -r --format="${WEEK_AGO} %(committerdate:short) %(refname:short)" --merged | sed 's/origin\///g' | egrep -v "(^\*|master|develop|release|HEAD)" | awk '$1 > $2' | awk '{ print $3 }' | xargs git push origin --delete 
-git prune
+if [ "$2" == "-r" ]
+then
+	echo "Pushing to remote"
+	WEEK_AGO=$(python -c "from datetime import date, timedelta; print(date.today()-timedelta(days=7))")
+	git branch -r --format="${WEEK_AGO} %(committerdate:short) %(refname:short)" --merged | sed 's/origin\///g' | egrep -v "(^\*|master|develop|release|HEAD)" | awk '$1 > $2' | awk '{ print $3 }' | xargs git push origin --delete 
+	git prune
+else
+	echo "Skipping remote"
+fi
